split files based on class type
